spp_primary: change type of ports_id to uint16_t
Type of ID and number of ports managed in primary is uint8 while it is
changed from uint8_t to uint16_t in DPDK 17.11.

This update is to change type of ID and number of ports to uint16_t.
It also includes updates for common functions called from primary.

+++ src/primary/args.c
@@ -66,7 +66,7 @@ usage(void)
  * on error.
  */
 int
-parse_app_args(uint8_t max_ports, int argc, char *argv[])
+parse_app_args(uint16_t max_ports, int argc, char *argv[])
 {
 	int option_index, opt;
 	char **argvopt = argv;
